# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is Documentation from the project assignment list application",
      "version": "1.0.0",
      "title": "Assignment List",
      "contact": {
        "email": "luis.gustavo.282000@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/users/roles": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Find Roles",
          "description": "Get All Roles",
          "operationId": "roles",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Role"
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        },
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Add new role",
          "description": "",
          "operationId": "addRole",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Role used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Role"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/users/roles/{id}": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Find Role by ID",
          "description": "Get Role By ID",
          "operationId": "getRoleById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Role",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Role"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        },
        "put": {
          "tags": [
            "Users"
          ],
          "summary": "Update Role",
          "description": "",
          "operationId": "updateRole",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Role",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Role used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Role"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        },
        "delete": {
          "tags": [
            "Users"
          ],
          "summary": "Delete Role by ID",
          "description": "Delete Role By ID",
          "operationId": "deleteRoleById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Role",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Role"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        }
      },
      "/users/roles/{id}/canApproven": {
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Approve In Role",
          "description": "",
          "operationId": "approve",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from Role",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Role used from the assing project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/PermitAproved"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/users/users": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Find Users",
          "description": "Get All Users",
          "operationId": "users",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        },
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Add new User",
          "description": "",
          "operationId": "addUser",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/users/users/{id}": {
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Find User by ID",
          "description": "Get User By ID",
          "operationId": "getUserById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from User",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        },
        "put": {
          "tags": [
            "Users"
          ],
          "summary": "Update User",
          "description": "",
          "operationId": "updateUser",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from User",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "User used from the insertion project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        },
        "delete": {
          "tags": [
            "Users"
          ],
          "summary": "Delete User by ID",
          "description": "Delete User By ID",
          "operationId": "deleteUserById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from User",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        }
      },
      "/users/users/{id}/canApproven": {
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Approve In User",
          "description": "",
          "operationId": "approve in user",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID from User",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "User used from the assing project",
              "required": true,
              "schema": {
                "$ref": "#/definitions/PermitAproved"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Role": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "canApprove": {
            "type": "boolean",
            "default": "false"
          },
          "canSuper": {
            "type": "boolean",
            "default": false
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "PermitAproved": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "client": {
              "type": "string"
            },
            "categories": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "canApprove": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }